Random item placement implemented
Repositioned the prefabs so the models sit on the ground when the parent objects are level with the ground. Added a script to randomly place the items at a given set of spawn points, and also keep track of when they're collected to assist with implementation of a win state.